pyproject.toml
[project]
name = "touche-rad"
version = "0.1.0"
authors = [
    { name = "Aaryan Potdar", email = "apotdar31@gatech.edu" },
    { name = "Anthony Miyaguchi", email = "acmiyaguchi@gatech.edu" },
    { name = "Conor Johnston", email = "cjohnston49@gatech.edu" }
]
description = "DS@GT CLEF Touche 2025 RAD"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26.4,<2.0",
    "pandas",
    "matplotlib",
    "scikit-learn",
    "tqdm",
    "pyspark>=3.4.0",
    "pyarrow",
    "luigi",
    "python-dotenv",
    "pyyaml",
    "openai",
    "chromadb",
    "sqlalchemy",
    "streamlit",
    "tensorzero",
    "torch",
    "sentence-transformers",
    "transitions",
    "jupyterlab",
    "ruff",
    "pre-commit",
    "pytest",
    "duckdb",
    "elasticsearch~=8.15.1",
    "fastapi[standard]>=0.115.9",
    "pydantic>=2.11.4",
]

[project.urls]
Homepage = "https://github.com/dsgt-kaggle-clef/touche-2025-rad"
Issues = "https://github.com/dsgt-kaggle-clef/touche-2025-rad/issues"


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["touche_rad*"]
namespaces = false
